+++ empymod/__init__.py
@@ -209,24 +209,9 @@ From the `FFTLog`-website:
 points.*
 
 FFTlog can be used for the Hankel as well as for the Fourier Transform, but
-currently `empymod` uses it only for the Fourier transform. By default it uses
-a condensed version of `pyfftlog` (`github.com/prisae/pyfftlog
-<https://github.com/prisae/pyfftlog>`_). You might get a small speed-up if you
-use `fftlog`, a simple python-wrapper of Fortran FFTLog code. However, it does
-not come pre-installed with `empymod`, as it has to be compiled on your system.
-You can download it from `github.com/prisae/fftlog
-<https://github.com/prisae/fftlog>`_ and install it into your python
-distribution (you need a Fortran compiler for this) by running the
-setup-script:
-
-.. code:: bash
-
-   python setup.py install
-
-I am currently working with `SciPy`-developers in this regard and hope that
-`FFTLog` will be included directly in `SciPy` in the not so distant future;
-this will then replace both of the above methods, the default `pyfftlog` or the
-optional `fftlog`, with `scipy.fftpack.fftlog`.
+currently `empymod` uses it only for the Fourier transform. It uses a
+simplified version of the python implementation of FFTLog, `pyfftlog`
+(`github.com/prisae/pyfftlog <https://github.com/prisae/pyfftlog>`_).